pyproject.toml
[project]
name = "tg-dev-platform"
version = "0.1.0"
description = "Autonomous Telegram-Driven Dev Platform"
authors = [
    { name = "Muvs", email = "your@email.com" },
]
readme = "docs/README.md"
requires-python = "^3.10"
dependencies = [
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "tg-dev-platform"
version = "0.1.0"
description = "Autonomous Telegram-Driven Dev Platform"
authors = ["Muvs <your@email.com>"]
readme = "docs/README.md"
packages = [
    { include = "services" },
    { include = "agents" },
]

[tool.poetry.dependencies]
python = ">=3.10, <=3.13"
pysqlite3-binary = "0.5.3.0"
fastapi = "^0.111.0"
python-telegram-bot = { extras = ["ext"], version = "^21.2" }
uvicorn = { extras = ["standard"], version = "^0.30.1" }
python-dotenv = "^1.0.1"
ruff = "^0.5.5"
black = "^24.4.2"
mypy = "^1.11.0"
pytest = "^8.3.2"
httpx = "^0.27.0"
crewai = { extras = ["tools"], version = "^0.134.0" }

[tool.poetry.group.dev.dependencies]
e2b = "^1.5.2"
pytest-cov = "^6.2.1"
pytest-asyncio = "^1.0.0"
